+++ CHANGELOG.md
@@ -10,6 +10,18 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Id type coercer can be automatically derived from the class constructor assuming such constructor exists : `constructor(raw: String)`
+- New notations for fields and for root objects :
+```kotlin
+query { }
+// is equivalent to
+query(object {}) { }
+
+query {
+    "fieldname" { -> "Yay" }
+    // is equivalent to
+    field("fieldname") { -> "Yay" }
+}
+```
 
 ### Fixed